% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric.R
\name{fabric}
\alias{fabric}
\title{Fabric.js In R}
\usage{
fabric(message = "fabric", width = NULL, height = NULL)
}
\description{
Unleash the amazing power of \href{http://fabricjs.com}{fabric.js}
  in R just to see all that we can do.
}

